+++ src/ngio/tables/backends/_json.py
@@ -34,21 +34,23 @@ class JsonTableBackend(AbstractTableBackend):
         """Whether the handler implements the polars protocol."""
         return True
 
-    def _get_table_group(self):
-        """Get the table group, creating it if it doesn't exist."""
+    def _get_table_handler(self):
+        """Get a handler for the table group, creating it if it doesn't exist."""
         try:
-            table_group = self._group_handler.get_group(path="table", create_mode=True)
+            table_handler = self._group_handler.get_handler(
+                path="table", create_mode=True
+            )
         except NgioError as e:
             raise NgioError(
                 "Could not get or create a 'table' group in the store "
                 f"{self._group_handler.store} path "
                 f"{self._group_handler.group.path}/table."
             ) from e
-        return table_group
+        return table_handler
 
     def load_as_pandas_df(self) -> DataFrame:
         """Load the table as a pandas DataFrame."""
-        table_dict = self._get_table_group().attrs.asdict()
+        table_dict = self._get_table_handler().load_attrs()
         data_frame = pd.DataFrame.from_dict(table_dict)
         data_frame = normalize_pandas_df(
             data_frame,
@@ -63,9 +65,7 @@ class JsonTableBackend(AbstractTableBackend):
 
     def _write_from_dict(self, table: dict) -> None:
         """Write the table from a dictionary to the store."""
-        table_group = self._get_table_group()
-        table_group.attrs.clear()
-        table_group.attrs.update(table)
+        self._get_table_handler().write_attrs(table, overwrite=True)
 
     def write_from_pandas(self, table: DataFrame) -> None:
         """Write the table from a pandas DataFrame."""